Revert "Split s3queue tests to be less heavy"
